Core: Extract ComponentMetaManager to Core
