+++ code/core/src/component-meta/index.ts
@@ -1,0 +1,12 @@
+export {
+  ComponentMetaManager,
+  isFileInDir,
+  parseTsconfigCommandLine,
+  sortTSConfigs,
+} from './ComponentMetaManager.ts';
+export type {
+  ComponentMetaProjectBase,
+  ComponentMetaProjectFactory,
+  FileChange,
+  FileChangeType,
+} from './types.ts';
